Create parseint function
Parse a string into a integer by looping through each character and
using the reverse index (first character highest index) as exponent to
calculate the power of 10 for the number. Multiply each number with it's
power of 10 and add it to a buffer.
Perform no sanity check on the string as that is already done in the
previous stringcheck function.

Signed-off-by: Sebastian Fricke <sebastian.fricke.linux@gmail.com>