Feature: calculate_equation function
Use the array from parse_equation to calculate an equation of variable
length, with the correct precedence of operaters.

The algorithm uses 3 precedence levels:
    - 3 (exponantiation/root)
    - 2 (multiplication/division)
    - 1 (addition/subtraction)
Search for all operations within the equation of a certain precedence
level, do the operation and replace the operation with the result.
Execute that action for all precedence levels and put the result into
RAX.

Signed-off-by: Sebastian Fricke <sebastian.fricke.linux@gmail.com>