Feature: Find the fractal part of a ASCII decimal
Scan through the string and search for a decimal point, return the
address to first character after the decimal point, which points to the
ASCII string of the fractal part.
Return -1 when more than one decimal point is found and return 0 when
there is no decimal point.

Function was tested on the following strings:
'123', '123.21', '123.21.21', '.123'

This is a preparation feature for the parse_float function.

Signed-off-by: Sebastian Fricke <sebastian.fricke.linux@gmail.com>