Feature: Split the number into subparts/zerocheck
Split the floating-point number into it's subparts:
sign, exponent & mantissa by applying bit masks to it.
Save the parts on the stack and return the address of the stack
location.
Check if the number is zero by comparing the mantissa & exponent with 0.

Preparation patch for the float->dec. string feature.

Signed-off-by: Sebastian Fricke <sebastian.fricke.linux@gmail.com>