Fix: Adjust incorrect masks
Adjust the masks used in `split_into_parts` to filter out the exponent
and mantissa correctly. Additionally store only the sign bit as 1 or 0.

Fix a failure within the zero check function, that read in a quad word
value out of the stack, where only a word was written. (Don't read
garbage values)

Signed-off-by: Sebastian Fricke <sebastian.fricke.linux@gmail.com>