Preparation: Compute interval of possible outputs
Preparation function for the print_float function, that implements the
'RYU' floating point to decimal string conversion algorithm.

Calculate the interval of possible ouputs by computing the halfway
points to the next bigger & lower floating point number.

The only special case are numbers with a mantissa of 0 and an exponent,
that is greater or equal to 153 (2^153). In this case the lower halfway
point is computed by subtracting 1 instead of 2.

Signed-off-by: Sebastian Fricke <sebastian.fricke.linux@gmail.com>