Refactor: Functions work directly with the stack
The algorithm needs a lot of different variables at different points,
this would imply to use many function parameter, to split the
functionality into smaller subparts.
To avoid this requirement, all core functions of the algorithm work directly on the stack and
get their inputs also directly from the stack.
That way the function is not reusable, but it improves
readability. The chain of functions works more like one
contiguous body of code.

Signed-off-by: Sebastian Fricke <sebastian.fricke.linux@gmail.com>