Feature: Conversion and Scale functionality
Implement the function to convert the output intervals into integer
using 64 bit arithmetic and a factor table. This is done by scaling up
the output interval with a factor from a table and shifting it to the
right by a specified value.

NOTE: Implements a simplified version, without checking for trailing
zeroes and removed digits. May provide incorrect results with ~4% of
possible inputs.

Signed-off-by: Sebastian Fricke <sebastian.fricke.linux@gmail.com>