Bug-Fix: Fix unexpected stack modification
The offset created initally for the set of variables on the stack was to
low. This caused a series of register pushs together with function entry
to set the stack pointer onto an address that was used by the struct.

Solve this problem by increasing the inital offset by 0x60 or 96.
(the size was chosen arbitrarily to reduce the chance of hitting this error again)

Signed-off-by: Sebastian Fricke <sebastian.fricke.linux@gmail.com>